+++ NEWS.md
@@ -3,6 +3,7 @@
 **New features**
 
 - support for `MASS::contr.sdif()` contrasts (#230)
+- support for `pscl::zeroinfl()` and `pscl::hurdle()` models (#232)
 
 **Fix**